Makefile: define THIRD_PARTY_SOURCES
After looking through the source files in compat/ and investigating the
files' content and/or its Git history, I've determined the list of files
that were copied from a third-party source. Place the names of these
files into the THIRD_PARTY_SOURCES variable in the Makefile.

In addition, add the sha1collisiondetection/ and sha1dc/ sources as well
since they are also imported from a third-party source.

In a future commit, this variable will be used to determine which files
are excluded when running the "coccicheck" target.

Signed-off-by: Denton Liu <liu.denton@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>